requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
kombu==5.5.3
numpy==2.2.6
oauthlib==3.2.2
openai==1.82.0
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pyjwt
pyparsing==3.2.3
pytesseract==0.3.13
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sendgrid-django
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
psycopg2-binary
azure-storage-blob
